+++ app/collections/[uuid]/page.tsx
@@ -73,7 +73,6 @@ export default async function Collection({
     ...searchParams,
     availableFilters: filteredAvailableFilters,
   };
-
   return (
     <PageLayout
       activePage="collection"
@@ -89,6 +88,10 @@ export default async function Collection({
         "collections",
         params.uuid
       )}
+      ga4Data={{
+        collection: collectionData.title,
+        division: collectionData.divisionTitle,
+      }}
     >
       <CollectionPage
         searchParams={updatedSearchParams}